refactor(cert): move csr assembly out of cert-client
